Restructure some code and get alter working
